update: Move news validation to her own file
