+++ controllers/newsController.js
@@ -1,29 +1,10 @@
-const { check, validationResult } = require('express-validator');
-
 const db = require('../models');
 const Entries = require('../models/entries');
 const httpCodes = require('../constants/constants');
 
 const DataTypes = db.Sequelize.DataTypes;
 
-const NEWS_ERROR_MESSAGES = {
-  TITLE_REQUIRED: 'La novedad necesita tener un título',
-  CONTENT_REQUIRED: 'La novedad necesita tener contenido',
-  IMAGE_REQUIRED: 'La novedad necesita tener una imagen',
-  IMAGE_TYPE: 'La imagen debe ser un archivo .png, .jpg o .jpeg',
-  CATEGORY_REQUIRED: 'La novedad necesita una categoría',
-};
-
 const postNews = async (req, res) => {
-  const errors = validationResult(req);
-
-  if (!errors.isEmpty()) {
-    return res.status(httpCodes.BAD_REQUEST).json({
-      ok: false,
-      errors: errors.array(),
-    });
-  }
-
   const categoryId = 1;
 
   const newsToInsert = {
@@ -52,16 +33,6 @@ const postNews = async (req, res) => {
   }
 };
 
-const validateNews = [
-  check('title').notEmpty().withMessage(NEWS_ERROR_MESSAGES.TITLE_REQUIRED),
-  check('content').notEmpty().withMessage(NEWS_ERROR_MESSAGES.CONTENT_REQUIRED),
-  check('image').notEmpty().withMessage(NEWS_ERROR_MESSAGES.IMAGE_REQUIRED),
-  check('category')
-    .notEmpty()
-    .withMessage(NEWS_ERROR_MESSAGES.CATEGORY_REQUIRED),
-];
-
 module.exports = {
   postNews,
-  validateNews,
 };
